+++ src/hooks/option/useCommandActions.ts
@@ -9,6 +9,14 @@ import {
   getAllCommandsFromFolder,
 } from '@/services/commandTree'
 
+/**
+ * Custom hook that provides actions for managing commands and folders in the command editor.
+ * Handles moving, organizing, and removing commands and folders with proper hierarchy management.
+ * 
+ * @param commandArray - React Hook Form field array for commands
+ * @param folderArray - React Hook Form field array for folders
+ * @returns Object containing command and folder management functions
+ */
 export const useCommandActions = (
   commandArray: ReturnType<
     typeof useFieldArray<CommandsSchemaType, 'commands'>
@@ -17,6 +25,11 @@ export const useCommandActions = (
 ) => {
   const tree = toCommandTree(commandArray.fields, folderArray.fields)
 
+  /**
+   * Moves a command from one position to another in the commands array
+   * @param from - Source index of the command
+   * @param to - Target index where the command should be moved
+   */
   const moveCommand = useCallback(
     (from: number, to: number) => {
       commandArray.move(from, to)
@@ -24,6 +37,12 @@ export const useCommandActions = (
     [commandArray],
   )
 
+  /**
+   * Moves a folder to become a child of another folder
+   * Prevents circular references by checking folder hierarchy
+   * @param sourceFolderId - ID of the folder to move
+   * @param targetFolderId - ID of the target parent folder (defaults to ROOT_FOLDER)
+   */
   const moveFolderToFolder = useCallback(
     (sourceFolderId: string, targetFolderId = ROOT_FOLDER) => {
       if (
@@ -50,6 +69,13 @@ export const useCommandActions = (
     [folderArray],
   )
 
+  /**
+   * Moves all commands within a folder to a new position in the commands array
+   * Handles both forward and backward drag operations to maintain correct order
+   * @param sourceFolderId - ID of the folder whose contents should be moved
+   * @param distIndex - Target index where the folder contents should be positioned
+   * @param firstChildIndex - Index of the first command in the folder
+   */
   const moveFolderContents = useCallback(
     (sourceFolderId: string, distIndex: number, firstChildIndex: number) => {
       const folderNode = findNodeInTree(tree, sourceFolderId)
@@ -74,6 +100,13 @@ export const useCommandActions = (
     [tree, commandArray, moveCommand],
   )
 
+  /**
+   * Moves a command to a specific folder and position
+   * Updates both the parent folder assignment and the position in the commands array
+   * @param commandId - ID of the command to move
+   * @param distIndex - Target index where the command should be positioned
+   * @param parentId - ID of the target parent folder (defaults to ROOT_FOLDER)
+   */
   const moveCommandToFolder = useCallback(
     (commandId: string, distIndex: number, parentId = ROOT_FOLDER) => {
       const idx = commandArray.fields.findIndex((c) => c.id === commandId)
@@ -91,6 +124,11 @@ export const useCommandActions = (
     [commandArray, moveCommand],
   )
 
+  /**
+   * Removes a command or folder from the arrays
+   * When removing a folder, moves all child commands and folders to the root folder
+   * @param id - ID of the command or folder to remove
+   */
   const commandRemove = useCallback(
     (id: string) => {
       const commandIndex = commandArray.fields.findIndex((c) => c.id === id)
