+++ src/hooks/option/useFavicon.tsx
